+++ samples/dSFMTdc/test_linearity.cpp
@@ -13,29 +13,29 @@ int main(int argc, char * argv[])
     }
     dSFMT_param params;
     char * para = argv[1];
-    params.mexp = strtoul(para, &para, 10);
+    params.mexp = static_cast<int>(strtoul(para, &para, 10));
     para++;
-    params.pos1 = strtoul(para, &para, 10);
+    params.pos1 = static_cast<int>(strtoul(para, &para, 10));
     para++;
-    params.sl1 = strtoul(para, &para, 10);
+    params.sl1 = static_cast<int>(strtoul(para, &para, 10));
     para++;
-    params.msk1 = strtoull(para, &para, 16);
+    params.msk1 = static_cast<uint64_t>(strtoull(para, &para, 16));
     para++;
-    params.msk2 = strtoull(para, &para, 16);
+    params.msk2 = static_cast<uint64_t>(strtoull(para, &para, 16));
     para++;
-    params.fix1 = strtoull(para, &para, 16);
+    params.fix1 = static_cast<uint64_t>(strtoull(para, &para, 16));
     para++;
-    params.fix2 = strtoull(para, &para, 16);
+    params.fix2 = static_cast<uint64_t>(strtoull(para, &para, 16));
     para++;
-    params.parity1 = strtoull(para, &para, 16);
+    params.parity1 = static_cast<uint64_t>(strtoull(para, &para, 16));
     para++;
-    params.parity2 = strtoull(para, &para, 16);
+    params.parity2 = static_cast<uint64_t>(strtoull(para, &para, 16));
     dSFMT sf(params);
     cout << sf.getParamString() << endl;
     bool success = true;
     for (int se = 1; se < 20; se++) {
         w128_t seedw;
-        seedw.u[0] = se;
+        seedw.u[0] = static_cast<uint32_t>(se);
         sf.seed(seedw);
         for (int i = 0; i < 2; i++) {
             for (int j = 2; j >= 1; j--) {
