+++ samples/sfmtdc/w128.hpp
@@ -214,11 +214,11 @@ namespace MTToolBox {
         if (x.u64[0] == 0) {
             int64_t y = (int64_t)x.u64[1];
             y = count_bit((uint64_t)(y & -y) - 1);
-            return 63 - y;
+            return 63 - static_cast<int>(y);
         } else {
             int64_t y = (int64_t)x.u64[0];
             y = count_bit((uint64_t)(y & -y) - 1);
-            return 63 - y + 64;
+            return 63 - static_cast<int>(y) + 64;
         }
     }
